Release 1.1.0

+++ resources/tools/bin/ballerina.bat
@@ -22,49 +22,86 @@ SetLocal EnableDelayedExpansion
 set CURRENT_PATH=%~sdp0
 set dist=false
 set update=false
+set build=false
+set RUN_BALLERINA=true
 set FILE_PATH=%CURRENT_PATH%..\distributions\ballerina-version
+set JAVA_CMD=java
 if "%1" == "dist" set dist=true
 if "%2" == "dist" set dist=true
 if "%1" == "update" set dist=true
+if "%2" == "update" set dist=true
+if "%1" == "build" set dist=true
 if "%1" == "update" set update=true
+if "%1" == "build" set build=true
 SetLocal EnableDelayedExpansion
+
+if exist %CURRENT_PATH%..\dependencies\jdk8u202-b08-jre (
+   set JAVA_CMD=%CURRENT_PATH%..\dependencies\jdk8u202-b08-jre\bin\java
+) else (
+    if not exist "%JAVA_HOME%" (
+        echo Compatible JRE 8 not found. Please follow the instructions in ^<BALLERINA_HOME^>\INSTALL.txt to install and setup Ballerina.
+        exit /b
+    )
+
+    "%JAVA_HOME%\bin\java" -version 2>&1 | findstr /r "1.8">NUL
+    if errorlevel 1 (
+        echo Compatible JRE 8 not found. Please follow the instructions in ^<BALLERINA_HOME^>\INSTALL.txt to install and setup Ballerina.
+        exit /b
+    )
+)
+
 if "%dist%" == "true" (
-   if exist %CURRENT_PATH%..\dependencies\jdk8u202-b08-jre (
-       %CURRENT_PATH%..\dependencies\jdk8u202-b08-jre\bin\java -jar %CURRENT_PATH%..\lib\ballerina-command-${ballerina.command.version}.jar %*
+   if not "%build%" == "true" (
+       set RUN_BALLERINA=false;
+   )
+   if "%build%" == "true" (
+        %JAVA_CMD% -jar %CURRENT_PATH%..\lib\ballerina-command-${ballerina.command.version}.jar build
    ) else (
-		java -jar %CURRENT_PATH%..\lib\ballerina-command-${ballerina.command.version}.jar %*
+        %JAVA_CMD% -jar %CURRENT_PATH%..\lib\ballerina-command-${ballerina.command.version}.jar %*
    )
    if "%update%" == "true" if exist  %CURRENT_PATH%..\ballerina-command-tmp (
         call %CURRENT_PATH%\..\ballerina-command-tmp\install.bat
+        if %errorlevel% neq 0 (
+            echo Update failed due to errors
+            rd /s /q %CURRENT_PATH%\..\ballerina-command-tmp
+            exit /b %errorlevel%
+        )
         rd /s /q %CURRENT_PATH%\..\ballerina-command-tmp
+        echo Update successfully completed
+        exit /b
    )
-) else (
-	set BALLERINA_HOME=
-	if exist %userprofile%\.ballerina\ballerina-version (
-	   set "FILE_PATH=%userprofile%\.ballerina\ballerina-version"
-	)
+)
 
-	SetLocal EnableDelayedExpansion
-	for /f %%a in (!FILE_PATH!) do (
-	  set BALLERINA_HOME=%%a
-	)
-	call %CURRENT_PATH%..\distributions\!BALLERINA_HOME!\bin\ballerina.bat %*
+if "%RUN_BALLERINA%" == "true" (
+    set BALLERINA_HOME=
+    for /f %%a in (%CURRENT_PATH%\..\distributions\ballerina-version) do (
+        set BALLERINA_HOME=%%a
+    )
+    if exist %userprofile%\.ballerina\ballerina-version (
+        set "FILE_PATH=%userprofile%\.ballerina\ballerina-version"
+    )
+
+    SetLocal EnableDelayedExpansion
+    for /f %%a in (!FILE_PATH!) do (
+        if exist %CURRENT_PATH%..\distributions\%%a (
+            set BALLERINA_HOME=%%a
+        )
+    )
+    call %CURRENT_PATH%..\distributions\!BALLERINA_HOME!\bin\ballerina.bat %*
 )
 set merge=false
 if "%1" == "help" (
 	if "%2" == "" set merge=true
 )
+if "%1" == "-h" set merge=true
+if "%1" == "--help" set merge=true
 if "%1" == "" set merge=true
 if "%1" == "version" set merge=true
 if "%1" == "-v" set merge=true
 if "%1" == "--version" set merge=true
 
 if "%merge%" == "true" (
-   if exist %CURRENT_PATH%..\dependencies\jdk8u202-b08-jre (
-       %CURRENT_PATH%..\dependencies\jdk8u202-b08-jre\bin\java -jar %CURRENT_PATH%..\lib\ballerina-command-${ballerina.command.version}.jar %*
-   ) else (
-		java -jar %CURRENT_PATH%..\lib\ballerina-command-${ballerina.command.version}.jar %*
-   )
+    %JAVA_CMD% -jar %CURRENT_PATH%..\lib\ballerina-command-${ballerina.command.version}.jar %*
 )
 
 exit /b
